requirements.txt
opencv-python
mediapipe
numpy
tensorflow
scikit-learn
keras
pandas
matplotlib
python-dotenv
twilio